update to avoid a null exception?
git-svn-id: http://svn.codehaus.org/geoserver/trunk@2892 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: d2bad8514a8a09ed9ece3a9b2a15125d38424faa [formerly 0c96ad76d90205426b1133467f482d9330b477c1]
Former-commit-id: 7bf3e5f5a1735eed287ab07f135808feaa83b7c2

+++ geoserver/src/org/vfny/geoserver/global/xml/XMLConfigWriter.java
@@ -39,7 +39,7 @@ import java.util.logging.Logger;
  * <p></p>
  *
  * @author dzwiers, Refractions Research, Inc.
- * @version $Id: XMLConfigWriter.java,v 1.31 2004/09/09 16:50:15 cholmesny Exp $
+ * @version $Id: XMLConfigWriter.java,v 1.32 2004/09/20 20:43:37 cholmesny Exp $
  */
 public class XMLConfigWriter {
     /** Used internally to create log information to detect errors. */
@@ -637,8 +637,10 @@ public class XMLConfigWriter {
 			if(fti == null){
 				//delete it
 				File[] t = fa[j].listFiles();
-				for(int x=0;x<t.length;x++)
+				if (t != null) {
+				    for(int x=0;x<t.length;x++)
 					t[x].delete();
+				}
 				fa[j].delete();
 			}
 		}
@@ -875,7 +877,7 @@ public class XMLConfigWriter {
  * <p></p>
  *
  * @author dzwiers, Refractions Research, Inc.
- * @version $Id: XMLConfigWriter.java,v 1.31 2004/09/09 16:50:15 cholmesny Exp $
+ * @version $Id: XMLConfigWriter.java,v 1.32 2004/09/20 20:43:37 cholmesny Exp $
  */
 class WriterUtils {
     /** Used internally to create log information to detect errors. */
